requirements.txt
click==8.1.7
pandas==1.5.3
torch==2.1.2
transformers==4.29.2
transformers-embedder==3.0.11
git+https://github.com/titipata/scipdf_parser
ollama
faiss_gpu
scikit-learn
transformers_embedder
